travis-build fix
